Refactor contact API handler into helper module
